requirements.txt
flask
flask-socketio
twilio
python-escpos
requests
pywin32; sys_platform == "win32"
pyusb; sys_platform != "win32"
eventlet; sys_platform != "win32"
